+++ src/output/itemRepository/ItemRepository.h
@@ -10,7 +10,10 @@ public:
     ItemData SetObtained(int64_t itemId);
     ItemData GetItem(int itemId);
     std::map<int64_t, ItemData> GetItems();
+    int AddEmblem();
+    int GetEmblemCount();
 
 private:
     std::map<int64_t, ItemData> _itemData;
+    int _emblemCount = 0;
 };